+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Ecuador region codes
 - Panama region codes
 
+### Fixed
+- (CYBRSOURCE-43) Added await to order Config changes
+
 ## [1.4.1] - 2022-08-17
 
 ### Fixed